Improved file structure. Added strudentjsonmanager. Bug fixes. Added more tests. Code refactored.
