+++ .gitignore
@@ -7,3 +7,4 @@ build/
 xcuserdata/
 *.xcuserstate
 *.DS_Store
+*.idea
