--- Lab2_class/StudentInfo/Contacts/contact.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-class Contact
-    def initialize(value)
-        @value = value
-    end
-    
-    def to_s
-        @value
-    end
-end
