--- Lab2_class/StudentInfo/Contacts/email_contact.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'contact'
-
-class EmailContact < Contact
-    def valid?
-        @value =~ /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i
-    end
-end
