--- Lab2_class/StudentInfo/Contacts/phone_contact.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'contact'
-
-class PhoneContact < Contact
-    def valid?
-        @value =~ /\A\+?\d{10}\z/
-    end
-end
