--- Lab2_class/StudentInfo/Contacts/telegram_contact.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'contact'
-
-class TelegramContact < Contact
-    def valid?
-        @value =~ /^@[A-Za-z0-9_]{5,32}$/
-    end
-end
