--- Lab2_class/StudentInfo/student.rb
@@ -1,92 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'student_validator'
-require_relative 'Contacts/contact_info'
-
-class Student
-    public
-    attr_reader :id, :name, :middle_name, :surname, :git, :contact_info
-    
-    def initialize(**kwargs)
-        @id = kwargs[:id]
-        @name = kwargs[:name]
-        @surname = kwargs[:surname]
-        @middle_name = kwargs[:middle_name]
-        @git = kwargs[:git]
-        
-        if kwargs[:contact_info]
-            @contact_info = ContactInfo.new(kwargs[:contact_info])
-        else
-            @contact_info = ContactInfo.from_string kwargs[:contact_info_string]
-        end
-        
-        validate
-    end
-    
-    def self.from_string(string)
-        values = string.split(',')
-        Student.new(
-                    id: values[0].strip,
-                    name: values[1].strip,
-                    surname: values[2].strip,
-                    middle_name: values[3].strip,
-                    git: values[4].strip,
-                    contact_info_string: values[5...].join(", ")
-                    )
-    end
-    
-    def to_s
-        "ID: #{id}, Surname: #{surname}, Name: #{name}, Middle Name: #{middle_name}, Contacts: #{contact_info.to_s}"
-    end
-    
-    def short_form
-        [id, name_initials, git, contact_info]
-    end
-    
-    def get_generic_info
-        "#{name_initials}, Git: #{@git}, Contacts: #{contact_info}"
-    end
-    
-    # Ex: Ivanov I.I.
-    def name_initials
-        "#{@surname} #{@name[0]}.#{@middle_name[0]}."
-    end
-    
-    # MARK: Setters
-    def id=(value)
-        @id = value
-        validate
-    end
-    
-    def name=(value)
-        @name = value
-        validate
-    end
-    
-    def middle_name=(value)
-        @middle_name = value
-        validate
-    end
-    
-    def surname=(value)
-        @surname = value
-        validate
-    end
-    
-    def git=(value)
-        @git = value
-        validate
-    end
-    
-    def contact_info=(value)
-        @contact_info = value
-        validate
-    end
-    
-    private
-    
-    def validate
-        validator = StudentValidator.new
-        validator.validate(student: self)
-    end
-end
