+++ Labs/Lab1/Lab1_start.rb
