+++ Labs/Lab2/Services/file_manager.rb
@@ -1,0 +1,9 @@
+module FileManager
+	attr_accessor :file_path, :file_extension
+
+	def load_data
+	end
+
+	def save_data
+	end
+end
