+++ Labs/Lab2/StudentInfo/Contacts/contact.rb
@@ -1,0 +1,9 @@
+class Contact
+	def initialize(value)
+		@value = value
+	end
+
+	def to_s
+		@value
+	end
+end
