+++ Labs/Lab2/StudentInfo/Contacts/email_contact.rb
@@ -1,0 +1,7 @@
+require_relative 'contact'
+
+class EmailContact < Contact
+	def valid?
+		@value =~ /\A[\w+\-.]+@[a-z\d\-]+(\.[a-z\d\-]+)*\.[a-z]+\z/i
+	end
+end
