+++ Labs/Lab2/StudentInfo/Contacts/phone_contact.rb
@@ -1,0 +1,7 @@
+require_relative 'contact'
+
+class PhoneContact < Contact
+	def valid?
+		@value =~ /\A\+?\d{10}\z/
+	end
+end
