+++ Labs/Lab2/StudentInfo/Contacts/telegram_contact.rb
@@ -1,0 +1,7 @@
+require_relative 'contact'
+
+class TelegramContact < Contact
+	def valid?
+		@value =~ /^@[A-Za-z0-9_]{5,32}$/
+	end
+end
