+++ Labs/Lab2/Test/JSON/output_json.json
@@ -1,0 +1,26 @@
+[
+  {
+    "id": 1,
+    "name": "Steve",
+    "middle_name": "Biden",
+    "surname": "Harvey",
+    "git": "https://github.com/Steve",
+    "contact_info": {
+      "phone": "+1234567890",
+      "email": "Steve@email.ru",
+      "telegram": ""
+    }
+  },
+  {
+    "id": 2,
+    "name": "Steve",
+    "middle_name": "Biden",
+    "surname": "Harvey",
+    "git": "https://github.com/Steve",
+    "contact_info": {
+      "phone": "+1234567890",
+      "email": "Steve@email.ru",
+      "telegram": ""
+    }
+  }
+]
