+++ Labs/Lab2/Test/TXT/output_txt.txt
@@ -1,0 +1,1 @@
+10, Harvey, Steve, Biden, https://github.com/Steve, +1234567890, Steve@email.ru,
